requirements.txt
Django==2.2.17
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-resized==0.3.11
djangorestframework==3.12.2
Pillow==8.1.0
pytz==2020.4
sqlparse==0.4.1
